+++ mosaic_utils/__init__.py
@@ -1,6 +1,6 @@
 #
 # This file is part of mosaic_utils package.
-# Copyright (C) 2025 HARMONIZE/INPE.
+# Copyright (C) 2026 HARMONIZE/INPE.
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by